requirements.txt
flet==0.21.2
python-docx==1.1.0
Pillow==10.1.0
reportlab==4.0.7
